refactor(AjaxResult): `com.team4.museum.util.ajax` 패키지로 이동
