+++ src/main/java/com/team4/museum/controller/action/member/LogoutAjaxAction.java
@@ -1,8 +1,8 @@
 package com.team4.museum.controller.action.member;
 
 import com.team4.museum.controller.action.AjaxAction;
-import com.team4.museum.util.AjaxResult;
 import com.team4.museum.util.UrlUtil;
+import com.team4.museum.util.ajax.AjaxResult;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
