+++ src/main/java/com/team4/museum/controller/action/member/MyPageFavoriteAjaxAction.java
@@ -7,7 +7,7 @@ import java.io.IOException;
 
 import com.team4.museum.controller.action.AjaxAction;
 import com.team4.museum.dao.FavoriteDao;
-import com.team4.museum.util.AjaxResult;
+import com.team4.museum.util.ajax.AjaxResult;
 import com.team4.museum.vo.MemberVO;
 
 import jakarta.servlet.http.HttpServletRequest;
