+++ src/main/java/com/team4/museum/controller/action/qna/QnaPwdCheckAjaxAction.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 import com.team4.museum.controller.action.AjaxAction;
 import com.team4.museum.dao.QnaDao;
-import com.team4.museum.util.AjaxResult;
+import com.team4.museum.util.ajax.AjaxResult;
 import com.team4.museum.vo.QnaVO;
 
 import jakarta.annotation.Nonnull;
